Move swiper code into own file
