+++ src/main.js
@@ -38,15 +38,14 @@ class OfferSelector {
         price: {unit: 'GBP'}
       }
     });
+  }
+
+  displayOffers(offers) {
     const nodes = document.getElementsByTagName(this.options.tag);
     if (nodes.length == 0) {
       throw Error(`Tag ${tag} not found in html`);
     }
-    this.parent = nodes[0];
-  }
-
-  displayOffers(offers) {
-    this.parent.innerHTML = '<offers></offers>';
+    nodes[0].innerHTML = '<offers></offers>';
     riot.mount('offers', {
       title: 'OPP Licence Offers',
       items: offers
@@ -54,7 +53,11 @@ class OfferSelector {
   }
 
   displayError(err) {
-    this.parent.innerHTML = '<error></error>';
+    const nodes = document.getElementsByTagName(this.options.tag);
+    if (nodes.length == 0) {
+      throw Error(`Tag ${tag} not found in html`);
+    }
+    nodes[0].innerHTML = '<error></error>';
     riot.mount('error', {
       error: err
     })
